% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{predict.ecoflex}
\alias{predict.ecoflex}
\title{Predict method for ecoflex models (generic dispatch)}
\usage{
\method{predict}{ecoflex}(object, newdata = NULL, type = "response", se.fit = FALSE, ...)
}
\arguments{
\item{object}{An ecoflex model object}

\item{newdata}{Optional new data frame for predictions}

\item{type}{Type of prediction ("response", "link", "prob")}

\item{se.fit}{If TRUE, return standard errors of predictions}

\item{...}{Additional arguments}
}
\description{
Predict method for ecoflex models (generic dispatch)
}
