% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{print.summary.ecoflex}
\alias{print.summary.ecoflex}
\title{Print method for summary.ecoflex objects}
\usage{
\method{print}{summary.ecoflex}(x, digits = 4, ...)
}
\arguments{
\item{x}{A summary.ecoflex object}

\item{digits}{Number of significant digits}

\item{...}{Additional arguments (ignored)}
}
\description{
Print method for summary.ecoflex objects
}
