% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdd_flex.R
\name{rdd_balance_test}
\alias{rdd_balance_test}
\title{Covariate Balance Test}
\usage{
rdd_balance_test(formula, data, cutoff = 0, covariates, ...)
}
\arguments{
\item{formula}{Formula with running variable}

\item{data}{Data frame}

\item{cutoff}{Cutoff point. Default: 0}

\item{covariates}{Character vector of covariate names to test}

\item{...}{Additional arguments}
}
\value{
Data frame with balance test results
}
\description{
Covariate Balance Test
}
