% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/did_flex.R
\name{plot.did_flex}
\alias{plot.did_flex}
\title{Event Study Plot for DiD}
\usage{
\method{plot}{did_flex}(x, type = "event_study", ci_level = 0.95, reference_period = -1, ...)
}
\description{
Event Study Plot for DiD
}
